Simplify middleware that serves frontend assets and redirects to index.html
