+++ src/middleware/mod.rs
@@ -7,6 +7,7 @@ mod prelude {
 pub use self::app::AppMiddleware;
 pub use self::current_user::CurrentUser;
 pub use self::debug::Debug;
+pub use self::ember_index_rewrite::EmberIndexRewrite;
 pub use self::head::Head;
 pub use self::security_headers::SecurityHeaders;
 pub use self::static_or_continue::StaticOrContinue;
@@ -14,6 +15,7 @@ pub use self::static_or_continue::StaticOrContinue;
 pub mod app;
 pub mod current_user;
 mod debug;
+mod ember_index_rewrite;
 mod head;
 mod security_headers;
 mod static_or_continue;
